fix android ndk r22 compability issues for android

+++ android/app/src/main/cpp/fftools_cmdutils.c
@@ -20,6 +20,9 @@
  */
 
 /*
+ * CHANGES 01.2021
+ * - NDK r22 incompatibility issues regarding INT64_MAX fixed
+ *
  * CHANGES 01.2020
  * - ffprobe support changes
  * - (optindex < argc) validation in parse_options() method updated with (optindex >= argc) check
@@ -358,7 +361,7 @@ static int write_option(void *optctx, const OptionDef *po, const char *opt,
     } else if (po->flags & OPT_BOOL || po->flags & OPT_INT) {
         *(int *)dst = parse_number_or_die(opt, arg, OPT_INT64, INT_MIN, INT_MAX);
     } else if (po->flags & OPT_INT64) {
-        *(int64_t *)dst = parse_number_or_die(opt, arg, OPT_INT64, INT64_MIN, INT64_MAX);
+        *(int64_t *)dst = parse_number_or_die(opt, arg, OPT_INT64, INT64_MIN, (double)INT64_MAX);
     } else if (po->flags & OPT_TIME) {
         *(int64_t *)dst = parse_time_or_die(opt, arg, 1);
     } else if (po->flags & OPT_FLOAT) {
